DEBUG init: trigger heartbeats as oneshot timeouts
This patch is an experiment to investigate sporadic timing artifacts of
the init test (the heartbeat test in particular) on some Qemu platforms.
The heartbeat mechansism is one of the very few users of the timer's
periodic-timeout feature. So the test may trigger a corner case in the
timer that has remained unnoticed so far. This is just speculation at
this point.

I merged the patch temporarily onto staging to see how the automated
tests respond to sidestepping the timer's periodic mode of operation.
The issue is not reproducable otherwise.

Issue #3079